requirements.txt
requests===2.28.1
beautifulsoup4===4.11.1
pandas===1.5.2
python-decouple==3.6
selenium===4.7.2
lxml===4.9.2
scholarly===1.7.9